Add modified time
